+++ app.go
@@ -7,12 +7,11 @@ import (
 	"strconv"
 
 	"github.com/TerraFactory/tilegenerator/database"
-	"github.com/TerraFactory/tilegenerator/geo"
-	"github.com/TerraFactory/tilegenerator/mapobjects"
+	"github.com/TerraFactory/tilegenerator/database/entities"
 	"github.com/TerraFactory/tilegenerator/settings"
-	"github.com/TerraFactory/tilegenerator/svg"
 	"github.com/fatih/color"
 	"github.com/gorilla/mux"
+	"github.com/TerraFactory/tilegenerator/tiles"
 )
 
 var db database.GeometryDB
@@ -35,94 +34,9 @@ func main() {
 	log.Fatal(http.ListenAndServe(fmt.Sprintf(":%s", conf.HTTPPort), router))
 }
 
-func createMapObject(dbObj geo.BaseGeometry) (*mapobjects.MapObject, error) {
-	switch dbObj.TypeID {
-	case 47:
-		return mapobjects.NewObject(
-			dbObj.ID,
-			dbObj.TypeID,
-			dbObj.Value,
-			`polyline, path, line {
-					stroke: black;
-					stroke-width: 1;
-					fill: none
-	       }`,
-			nil)
-	case 74:
-		return mapobjects.NewObject(
-			dbObj.ID,
-			dbObj.TypeID,
-			dbObj.Value,
-			`line {
-			fill: none;
-			stroke: red;
-			}`,
-			nil)
-	case 59: //ГЦСС
-		return mapobjects.NewObject(
-			dbObj.ID,
-			dbObj.TypeID,
-			dbObj.Value,
-			`line {
-			fill: none;
-			stroke: red;
-			}`,
-			dbObj.Attrs)
-	default:
-		if dbObj.TypeID >= 149 && dbObj.TypeID <= 165 {
-			return mapobjects.NewObject(
-				dbObj.ID,
-				dbObj.TypeID,
-				dbObj.Value,
-				`line {
-			fill: none;
-			stroke: red;
-			}`,
-				dbObj.Attrs)
-		}
-
-		return nil, fmt.Errorf("Unexpected object type: %v", dbObj)
-	}
-}
 
 func getTile(writer http.ResponseWriter, req *http.Request) {
-	var objects []mapobjects.MapObject
-	var ids []int
-
-	results, objErr := db.GetAllPatrollingAreas()
-	if objErr != nil {
-		fmt.Println(objErr.Error())
-		return
-	}
-	for _, obj := range results {
-		ids = append(ids, obj.ID)
-	}
-
-	attrs, attrErr := db.GetAllAttributes(ids)
-	if attrErr != nil {
-		fmt.Println(attrErr.Error())
-		return
-	}
-
-	for _, r := range results {
-		var objAttrs []geo.BaseAttribute
-		for _, attr := range attrs {
-			if attr.ObjectID == r.ID {
-				objAttrs = append(objAttrs, attr)
-			}
-		}
-
-		r.Attrs = objAttrs
-
-		obj, err := createMapObject(r)
-		if err == nil {
-			objects = append(objects, *obj)
-		} else {
-			fmt.Println(err.Error())
-			writer.WriteHeader(400)
-			return
-		}
-	}
+	var objects []entities.MapObject
 
 	vars := mux.Vars(req)
 	x, errX := strconv.Atoi(vars["x"])
@@ -133,7 +47,7 @@ func getTile(writer http.ResponseWriter, req *http.Request) {
 		return
 	}
 
-	tile := mapobjects.NewTile(x, y, z)
+	tile := tiles.NewTile(x, y, z)
 	writer.Header().Set("Content-Type", "image/svg+xml")
-	svg.RenderTile(tile, &objects, writer)
+	tiles.RenderTile(tile, &objects, writer)
 }
