+++ tiles/rendering.go
@@ -1,16 +1,16 @@
-package svg
+package tiles
 
 import (
 	"io"
 
 	"github.com/TerraFactory/svgo"
-	"github.com/TerraFactory/tilegenerator/mapobjects"
+	"github.com/TerraFactory/tilegenerator/database/entities"
 )
 
 // RenderTile takes a tile struct, map objects and then draws these objects on the tile
-func RenderTile(tile *mapobjects.Tile, objects *[]mapobjects.MapObject, writer io.Writer) {
+func RenderTile(tile *Tile, objects *[]entities.MapObject, writer io.Writer) {
 	canvas := svg.New(writer)
-	canvas.Start(mapobjects.TileSize, mapobjects.TileSize)
+	canvas.Start(TileSize, TileSize)
 	//for _, _ := range *objects {
 	//}
 	canvas.End()
